requirements.txt
Django==1.7.2
django-picklefield==0.3.1
emoji==0.1.0
oauthlib==0.7.2
psycopg2==2.5.4
requests==2.4.3
requests-oauthlib==0.4.1
six==1.7.3
tweepy==3.1.0
wsgiref==0.1.2
